Add files via upload
Diagnosing COVID-19 using AI-based medical image analyses

Problem Statement:
Analyze suspected Covid-19 Patient’s Chest X-Rays and confirm if Positive for Virus infection.
Radiologists have noticed the coronavirus disease i.e. COVID-19 pandemic unfold in recent months. Radiology literature suggests a pivotal role for X-RAY as X-Ray findings in Covid-19 patient has pneumonia, and hence medical imaging has high sensitivity for diagnosis of COVID-19. The objective of this statement is to figure out usage of AI in medical imaging modalities for fast and accurate diagnosis of COVID-19.

Goal:
Diagnosing or detecting patients infected with COVID-19 using chest X-Ray images.

Use Case Details:
Doctors need to upload pictures of an individual’s X-ray or click the picture of a patient’s x-ray. The App needs to classify whether the patient has any sign of pneumonia, classify whether it is induced by bacteria or Covid-19 Virus and also measure severity of the infection.

The software should help in accurate preliminary screening which can be supplemented by clinical testing for those tested positive for the virus. The pneumonia caused by Covid-19 is severe than other bacteria as it affects lungs completely than smaller portions of lungs in other cases. The software should analyze bilateral opacity, pattern of fluid build-up in lungs and nature of clump or clot if any.

The main application of this software is that it should be able to process thousands of x-ray images in a matter of minutes and classify a patient according to the level of coronal infection so to help hospital people to decide which patient needs to be admitted to ICU or put on ventilator first.